update balance outside main loop
